requirements.txt
Django==5.2.8
Pillow==12.0
django-allauth
django-crispy-forms
python-dotenv
django-filter
djangorestframework
oauthlib
requests-oauthlib
